+++ src/client/AvanzaClient.ts
@@ -12,7 +12,7 @@ export const AVANZA_URL = "https://www.avanza.se";
 
 export interface AvanzaClientOptions {
   baseUrl?: string;
-  fetch?: Fetcher;
+  fetch: Fetcher;
 }
 
 export class AvanzaClient {
@@ -24,7 +24,7 @@ export class AvanzaClient {
 
   constructor(options?: AvanzaClientOptions) {
     this.baseUrl = options?.baseUrl || AVANZA_URL;
-    this.fetch = options?.fetch || require("node-fetch");
+    this.fetch = options.fetch;
     this.auth = new AuthClient(this);
     this.account = new AccountClient(this);
   }