+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- Made fetch mandatory to fix require crashing if node-fetch is not installed
+
 ## 0.2.0 - 2022-11-20
+
 ### Added
+
 - Changelog